requirements.txt
flask
requests
beautifulsoup4
html5lib
python-telegram-bot==13.15
apscheduler
python-dotenv